+++ subprojects/groovy-macro/src/main/java/org/codehaus/groovy/macro/transform/MacroTransformation.java
@@ -15,56 +15,25 @@
  */
 package org.codehaus.groovy.macro.transform;
 
-import org.codehaus.groovy.ast.ASTNode;
-import org.codehaus.groovy.ast.ClassCodeVisitorSupport;
-import org.codehaus.groovy.ast.ClassNode;
-import org.codehaus.groovy.ast.ModuleNode;
-import org.codehaus.groovy.ast.expr.MethodCallExpression;
+import org.codehaus.groovy.ast.*;
 import org.codehaus.groovy.control.CompilePhase;
 import org.codehaus.groovy.control.SourceUnit;
-import org.codehaus.groovy.transform.ASTTransformation;
 import org.codehaus.groovy.transform.GroovyASTTransformation;
 
 /**
  *
  * @author Sergei Egorov <bsideup@gmail.com>
  */
+
 @GroovyASTTransformation(phase = CompilePhase.CONVERSION)
-public class MacroTransformation extends ClassCodeVisitorSupport implements ASTTransformation {
+public class MacroTransformation extends MethodCallTransformation {
 
     public static final String DOLLAR_VALUE = "$v";
     public static final String MACRO_METHOD = "macro";
 
-    SourceUnit source;
-    
-    MacroInvocationTrap transformer;
-
-    public void visit(ASTNode[] nodes, SourceUnit source) {
-        this.source = source;
-
-        transformer = new MacroInvocationTrap(source.getSource(), source);
-
-        for(ASTNode node : nodes) {
-            if(node instanceof ClassNode) {
-                visitClass((ClassNode) node);
-            } else if(node instanceof ModuleNode) {
-                ModuleNode moduleNode = (ModuleNode) node;
-                for (ClassNode classNode : moduleNode.getClasses()) {
-                    visitClass(classNode);
-                }
-            }
-        }
-    }
-
     @Override
-    public void visitMethodCallExpression(MethodCallExpression call) {
-        transformer.visitMethodCallExpression(call);
-        super.visitMethodCallExpression(call);
-    }
-
-    @Override
-    protected SourceUnit getSourceUnit() {
-        return source;
+    protected GroovyCodeVisitor getTransformer(ASTNode[] nodes, SourceUnit sourceUnit) {
+        return new MacroInvocationTrap(sourceUnit.getSource(), sourceUnit);
     }
 }
